requirements.txt
aiocron==1.8
aiofiles==23.2.1
aiogram==3.2.0
aiohttp==3.9.0
aioredis==1.3.0
aioschedule==0.5.2
aiosignal==1.3.1
annotated-types==0.6.0
APScheduler==3.10.4
async-timeout==4.0.3
attrs==23.1.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
contourpy==1.2.0
croniter==2.0.1
cycler==0.12.1
fonttools==4.48.1
frozenlist==1.4.0
google-api-core==2.15.0
google-api-python-client==2.111.0
google-auth==2.25.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.62.0
greenlet==3.0.3
hiredis==2.3.2
httplib2==0.22.0
idna==3.6
kiwisolver==1.4.5
magic-filter==1.0.12
matplotlib==3.8.2
multidict==6.0.4
numpy==1.26.4
oauthlib==3.2.2
pandas==2.2.0
pillow==10.2.0
plotly==5.18.0
protobuf==4.25.1
psycopg2==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.5.2
pydantic-settings==2.1.0
pydantic_core==2.14.5
pyparsing==3.1.1
python-dotenv==1.0.0
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
seaborn==0.13.2
SQLAlchemy==2.0.25
style==1.1.0
tenacity==8.2.3
typing_extensions==4.8.0
tzdata==2023.4
tzlocal==5.2
update==0.0.1
uritemplate==4.1.1
urllib3==2.1.0
yarl==1.9.3
